Apply random color to square on mouse over
Background color of a grid square set to a random color on mouseover. Every time a grid square is moused over, 10% black is added to the color. On the 10th mouseover, the grid square becomes black.

Mouseover count and original color stored using data attributes. These
data attributes are used for the calculation of adding 10% black to the
square's color. randomRGB() was split into randomColorIntensity() and getRGB() to accomodate adding 10% black to an existing color.