Handle non-numerical inputs for grid dimensions
If a user cancels, enters an empty string, or enters a non-numerical
input when prompted to enter the number of squares per side for the new
grid, the program will use the previous valid squares per side.
Previously the grid would not be created if either erroneous inputs were
used.